babel: extract components, elements
